+++ packages/app-security/src/admin/plugins/installation.tsx
@@ -5,6 +5,7 @@ import { Form } from "@webiny/form";
 import { i18n } from "@webiny/app/i18n";
 import { ButtonPrimary } from "@webiny/ui/Button";
 import { Input } from "@webiny/ui/Input";
+import { Checkbox } from "@webiny/ui/Checkbox";
 import { Alert } from "@webiny/ui/Alert";
 import { Typography } from "@webiny/ui/Typography";
 import { Grid, Cell } from "@webiny/ui/Grid";
@@ -66,7 +67,7 @@ const Install = ({ onInstalled }) => {
     const [authUserMessage, setAuthUserMessage] = useState(null);
     const [error, setError] = useState(null);
 
-    const onSubmit = useCallback(async form => {
+    const onSubmit = useCallback(async ({ subscribed, ...form}) => {
         setLoading(true);
         setError(null);
         if (typeof auth.onSubmit === "function") {
@@ -92,8 +93,28 @@ const Install = ({ onInstalled }) => {
             return;
         }
 
+        if(subscribed) {
+            try {
+                await fetch (
+                    'https://app.mailerlite.com/webforms/submit/g9f1i1?fields%5Bemail%5D=' +
+                        encodeURIComponent(form.email) +
+                    '&ml-submit=1&ajax=1',
+                    {
+                    method: 'GET',
+                    headers: {
+                        'Content-Type': 'application/x-www-form-urlencoded',
+                    },
+                    }
+                )               
+            } catch (err) {
+                setError("Unable to subscribe you to the newsletter "+ err);
+            }
+        }
+
         onInstalled();
     }, []);
+    
+    const privacyPolicyLink = <a href="https://www.webiny.com/privacy-policy">privacy policy</a>;
 
     return (
         <Form onSubmit={onSubmit} submitOnEnter>
@@ -156,6 +177,15 @@ const Install = ({ onInstalled }) => {
                                     )
                                 }
                             })}
+                            {!authUserMessage && (
+                                <Grid>
+                                    <Cell span={12}>
+                                        <Bind name="subscribed">
+                                            <Checkbox label={<span>I want to receive updates on product improvements and new features. Doing so I accept Webiny's {privacyPolicyLink}.</span> }/>
+                                        </Bind>
+                                    </Cell>
+                                </Grid>
+                            )}
                     </SimpleFormContent>
                     <SimpleFormFooter>
                         {!authUserMessage && (