+++ packages/app-cookie-policy/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 4.0.0-beta.9 (2020-06-02)
+
+**Note:** Version bump only for package @webiny/app-cookie-policy
+
+
+
+
+
+# 4.0.0-beta.8 (2020-06-02)
+
+
+### Features
+
+* ✨  Remove unneeded text below password. Add link to checkbox ([39687f4](https://github.com/webiny/webiny-js/commit/39687f42f17c0066c681c16745ab9c37d5759f08))
+
+
+
+
+
+# 4.0.0-beta.7 (2020-06-01)
+
+**Note:** Version bump only for package @webiny/app-cookie-policy
+
+
+
+
+
 # [4.0.0-beta.6](https://github.com/webiny/webiny-js/compare/v4.0.0-beta.5...v4.0.0-beta.6) (2020-06-01)
 
 **Note:** Version bump only for package @webiny/app-cookie-policy